Fignolage : - Organisation AdminController - Ajustement CSS : suppression des classes btn inutilisées + bootstrap - Suppression du border-radius de l'avatar - Suppression de certain margin pour les tableaux - Modification des scrollbar : overflow en auto - Suppression des use inutilisés dans les Controller - Changement du title de la page d'accueil - Ajout du lien vers la fiche de l'organisateur d'une sortie - Bouton Créer une sortie sur la page d'accueil - Changement de certains noms de routes accessibles uniquement depuis le dashboard - Nav bar : mise en place des liens vers Ville et Campus + changement de la position de "Se déconnecter"
