+++ src/Controller/LoginController.php
@@ -29,4 +29,5 @@ class LoginController extends AbstractController
     {
         throw new \LogicException('This method can be blank - it will be intercepted by the logout key on your firewall.');
     }
+
 }
