+++ templates/admin/inscription.html.twig
@@ -14,7 +14,9 @@
     {{ form_start(registrationForm) }}
 
         {{ form_widget(registrationForm)}}
-        <button type="submit" class="submit">Enregistrer</button>
+
+            <button type="submit" class="btn btn-success">Enregistrer</button>
+
     {{ form_end(registrationForm) }}
 
     </div>
