+++ templates/admin/list.html.twig
@@ -7,7 +7,6 @@
     <h4 class="adminTitle">Liste des utilisateurs</h4>
 
     <div class="container">
-
     <table id="list">
         <tr>
             <td>SUPPRESSION</td>
@@ -39,8 +38,8 @@
 
     </div>
     
-    <section class="container">
-        <a href="{{ path('admin_dashboard') }}" title="retour"><button class="retour">Retour</button></a>
+    <section class="container padding">
+        <a href="{{ path('admin_dashboard') }}" title="retour"><button class="btn btn-secondary">Retour</button></a>
     </section>
 
 {% endblock %}
