+++ templates/base.html.twig
@@ -31,10 +31,8 @@
     {% block javascripts %}{% endblock %}
 
     {% block body %}
-
-
-
     {% endblock %}
+
 </main>
 
 <footer>
