+++ templates/profile/fiche.html.twig
@@ -29,8 +29,8 @@
         </section>
     </div>
 
-    <div class="button_retour">
-    <a href="{{ path('main_accueil') }}" title="Accueil"><button class="retour">Retour</button></a>
+    <div class="container padding">
+    <a href="{{ path('main_accueil') }}" title="Accueil"><button class="btn btn-secondary">Retour</button></a>
     </div>
