+++ templates/security/login.html.twig
@@ -29,8 +29,7 @@
            </label>
         </div>
 
-
-        <button class="submit" type="submit">
+        <button class="btn btn-success" type="submit">
             Connexion
         </button>
